location transfer: ensure move state is recomputed
When we split a move, the state of the remaining part may remain in
"waiting" if we do not call "move._recompute_state()".

In the order of events:

* the move is split, the remaining part becomes a "waiting" move
* the existing remaining move lines are directly assigned to the
  "waiting move"
* when _action_assign() is called, as the move is "waiting", it checks
  if it needs reservation, but as it already has all the move lines,
  it does not changes the move to "assigned"

By calling "move._recompute_state()", if the reservation is already
full, the state becomes "assigned" (and the move is ignored by
"_action_assign()").