[FIX] shopfloor, checkout: ability to recover/restart
If the user starts to process a line, and for whatever reason he
stops there and restarts the scenario from the beginning, he should
still be able to find/recover the previous line.

For that purpose we are using the `shopfloor_user_id` field on move
lines (already used to manage currently processed lines in the Zone
picking and Location content transfer scenarios).